Relocate filesToStage into RunnerContext
